pyproject.toml
[project]
name = "quickcall-voiceover"
version = "0.1.0"
description = "Utility tool for creating voice-over audio assets for QuickCall videos"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [
    { name = "QuickCall", email = "hello@quickcall.dev" }
]
keywords = ["voiceover", "tts", "piper", "audio", "quickcall"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
]
dependencies = [
    "piper-tts>=1.3.0",
]

[project.urls]
Homepage = "https://github.com/quickcall-dev/quickcall-voiceover"
Repository = "https://github.com/quickcall-dev/quickcall-voiceover"
Issues = "https://github.com/quickcall-dev/quickcall-voiceover/issues"

[project.scripts]
quickcall-voiceover = "quickcall_voiceover.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["quickcall_voiceover"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
